Added covariates to model parametrizations
Three new models added based on m1 parametrization, including: chl, depth and chl + depth. Model selection based on log-likelihoods reveals that the performance of the model improves when chl is included. Depth effect not significant. Coefficient extraction and predictions chunks updated accordingly. A bug related to the replacement of year_ct with yday_ct was fixed